Fixes bug where grouping events did not always return the the most chronologically earliest matching date.
This is fixed by using an INNER join rather than LEFT join. This however effectivley reverted the fix for
inner join with a result set that did not include that events ID.

The fix is to check for any date-based where conditinos. If none are present that it is resonable to return
events that do not have any dates. Typically this might only be expected on the Events admin page.

Ref #367