We should calculate an event's dates in the timezone in which it is defined.
This is because the Daylight Savings boundaries may be different in the two timezones. For example consider
a monthly event in Perth, Australia whose occurrences are:

    2nd Sept 2016 5am, 2nd Oct 2016 5am, 2nd Nov 2016 5am

If we converted the event's first start date to Europe/London and recalculated we get:

    1st Sept 2016 10pm, 1st Oct 2016 10pm, 1st Nov 2016 10pm

But if we converted each date individually we notice taht the correct DateTimes in Europe/London are

    1st Sept 2016 10pm, 1st Oct 2016 10pm, 1st Nov 2016 9pm

(note last datetime) - as the last date falls outside of DST in Europe/London.

Ref #377.