Unit test describing existing behaviour of converting a 'foreign' event into the blog timezone.
Since events are converted to and consequently 'defined' in the blog timezone. It is sometimes
necessary to adjust the definition of the event so that it is accurate when converting it across
timezones.

This introduces some edge cases (see #377) and future solution might store an event's initial timezone
and define the events in their original timezone.

Ref #377.