Improve tests so that the number of comments/plugins are regexed
Currently we expect 0 plugins and 0 comments in the admin menu. This might
not always be the case for legitimate reasons. (I.e. running an older version
of Event Organiser in the test) and we generally do not care what the number
is.

This commit replaces a strict comparison when a row-by-row regex comparison.

It leaves open the possibility of expecting a particular number.