Fix duration of long-running occurrences incorrectly calculated.
Previously the duration of an occurence had been calculated in years,
months and days (and time). But this means an event with a duration
of over a month will potentially have occurrences of differing durations.

For example an occurrence running from 17th January 2018 to 7 March lasts
1 month & 4 days, or 7 weeks. If that occurrence is repeated on the 13th June,
it will last only 48 days because that is the equivalent of 1 month and 4 days.

This variation is undesirable and we should use the more exact measurement of
duration: days (and time) only.