If an event has occurrences which have changed time, do not reset that time when editing the event ocurrences UNLESS the event's initial start/end datetime has changed.
This will allow users to extend the occurrences of the event without loosing their changes
to individual occurrences.

In the future this check *may* become more strict: i.e. any change to the recurrence pattern
beyond extending it or shortening it would cause a reset of all the occurrence times.