Fix infinite loop bug if maps.length > locations.length + 1
This might happen, for instance, if Yoast's SEO plug-in is installed.
maps.length might 6 (although only the last instance is rendered), and
locations.length = 1. Because the variable i is being re-used, the
inner loop then effectively reset it to 1, the outer for loop
increments it by 1 (so i=2), but then i < maps.length, so we never
escape the outer for loop.