`event.js`: Fix test for matching the start-date datepicker
The element ID of the start-date datepicker in the event editor
has changed: it is not `from_date` anymore, but `eo-start-date`.

However, since the role of each datepicker is now accessible
using `$(element).data('eo-datepicker')`, matching this string
should provide a more robust test for figuring out if we are
currently handling events for a particular datepicker.